NNM-82 뷰에 지도 추가
