+++ Nanuming/View/MapView.swift
@@ -5,38 +5,34 @@
 //  Created by 가은 on 1/24/24.
 //
 
-import CoreLocation
+import GoogleMaps
 import SwiftUI
 
 struct MapView: View {
-
+    @ObservedObject var mapVM: MapViewModel
+    
     var body: some View {
-        Text("mapView")
-            .onAppear {
-                requestLocationAuthorization()
-            }
-    }
-    
-}
-
-// 위치 권한 확인
-func requestLocationAuthorization() {
-    let locationManager = CLLocationManager()
-    
-    // 위치 사용 권한 거부된 상태
-    if locationManager.authorizationStatus == .denied {
-        DispatchQueue.main.async {
-            UIApplication.shared.open(URL(string: UIApplication.openSettingsURLString)!)
-        }
-    }
-    // 위치 사용 권한 대기 상태
-    else if locationManager.authorizationStatus == .restricted || locationManager.authorizationStatus == .notDetermined {
-        // 권한 요청
-        locationManager.requestWhenInUseAuthorization()
+        GoogleMapView(mapVM: mapVM)
     }
 }
 
+struct GoogleMapView: UIViewRepresentable {
+    @ObservedObject var mapVM: MapViewModel
+    
+    func makeUIView(context: Context) -> GMSMapView {
+        mapVM.requestLocationAuthorization()
+        
+        let mapView = GMSMapView()
+        mapView.isMyLocationEnabled = true
+        mapView.settings.myLocationButton = true
+        
+        let camera = GMSCameraPosition.camera(withLatitude: 37.5665, longitude: 126.9780, zoom: 12)
+        mapView.camera = camera
+        
+        return mapView
+    }
 
-#Preview {
-    MapView()
+    func updateUIView(_ uiView: GMSMapView, context: Context) {
+        print("update")
+    }
 }
